move some packages around
