+++ src/main/java/com/github/thedeathlycow/frostiful/Frostiful.java
@@ -13,8 +13,8 @@ import com.github.thedeathlycow.frostiful.server.command.WindCommand;
 import com.github.thedeathlycow.frostiful.server.network.PointWindSpawnPacket;
 import com.github.thedeathlycow.frostiful.sound.FSoundEvents;
 import com.github.thedeathlycow.frostiful.survival.*;
-import com.github.thedeathlycow.frostiful.world.gen.feature.FFeatures;
-import com.github.thedeathlycow.frostiful.world.gen.feature.FPlacedFeatures;
+import com.github.thedeathlycow.frostiful.server.world.gen.feature.FFeatures;
+import com.github.thedeathlycow.frostiful.server.world.gen.feature.FPlacedFeatures;
 import com.github.thedeathlycow.thermoo.api.temperature.event.EnvironmentControllerInitializeEvent;
 import com.github.thedeathlycow.thermoo.api.temperature.event.PlayerEnvironmentEvents;
 import me.shedaniel.autoconfig.AutoConfig;
@@ -24,7 +24,6 @@ import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
 import net.fabricmc.fabric.api.loot.v3.LootTableEvents;
 import net.fabricmc.fabric.api.networking.v1.PayloadTypeRegistry;
-import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
 import net.fabricmc.fabric.api.util.TriState;
 import net.minecraft.util.Identifier;
 import org.jetbrains.annotations.Contract;
