+++ src/main/java/com/github/thedeathlycow/frostiful/entity/RootedEntity.java
@@ -2,7 +2,7 @@ package com.github.thedeathlycow.frostiful.entity;
 
 import com.github.thedeathlycow.frostiful.Frostiful;
 import com.github.thedeathlycow.frostiful.config.FrostifulConfig;
-import com.github.thedeathlycow.frostiful.enchantment.IceBreakerEnchantment;
+import com.github.thedeathlycow.frostiful.item.enchantment.IceBreakerEnchantment;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.MovementType;
