+++ src/main/java/com/github/thedeathlycow/frostiful/server/world/gen/feature/coveredrock/CoveredRockFeature.java
@@ -1,4 +1,4 @@
-package com.github.thedeathlycow.frostiful.world.gen.feature.coveredrock;
+package com.github.thedeathlycow.frostiful.server.world.gen.feature.coveredrock;
 
 import com.github.thedeathlycow.frostiful.registry.tag.FBlockTags;
 import com.mojang.serialization.Codec;
